Move `ZarrOpeningSettings`, `ZarrOpenBehavior`, and `ZarrReaderBackend` to `open.options` package
