+++ src/main/java/sc/fiji/ome/zarr/open/options/ZarrReaderBackend.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.settings;
+package sc.fiji.ome.zarr.open.options;
 
 import java.util.NoSuchElementException;
