+++ src/test/java/sc/fiji/ome/zarr/pyramid/Pyramidal5DImageDataOpenTest.java
@@ -43,8 +43,8 @@ import org.scijava.prefs.PrefService;
 import sc.fiji.ome.zarr.pyramid.backend.n5.N5BackedPyramidal5DImageDataTest;
 import sc.fiji.ome.zarr.pyramid.backend.zarrjava.ZarrJavaBackedPyramidal5DImageDataTest;
 import sc.fiji.ome.zarr.pyramid.exceptions.NoMatchingResolutionException;
-import sc.fiji.ome.zarr.settings.ZarrOpeningSettings;
-import sc.fiji.ome.zarr.settings.ZarrReaderBackend;
+import sc.fiji.ome.zarr.open.options.ZarrOpeningSettings;
+import sc.fiji.ome.zarr.open.options.ZarrReaderBackend;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
 /**
